+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 80.0.0
 
 ### Enhancements
+- **Adverse Event Changes**
+  - Adverse event admins can delete closed adverse events
 - **Gem Changes**
   - Update to rails 6.0.0
   - Update to devise 4.7.1